requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_tflow_190_select=0.0.1=gpu
absl-py=0.3.0=py36_0
appdirs=1.4.3=py36h28b3542_0
asn1crypto=0.24.0=py36_0
astor=0.6.2=py36_1
attrs=18.1.0=py36_0
automat=0.7.0=py36_0
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=2.1.3=py36_0
ca-certificates=2018.03.07=0
certifi=2018.4.16=py36_0
cffi=1.11.5=py36h9745a5d_0
constantly=15.1.0=py36h28b3542_0
cudatoolkit=9.0=h13b8566_0
cudnn=7.1.2=cuda9.0_0
cupti=9.0.176=0
cython=0.28.4=py36hf484d3e_0
decorator=4.3.0=py36_0
entrypoints=0.2.3=py36_2
freetype=2.9.1=h8a8886c_0
gast=0.2.0=py36_0
gmp=6.1.2=h6c8ec71_1
grpcio=1.12.1=py36hdbcaa40_0
html5lib=1.0.1=py36_0
hyperlink=18.0.0=py36_0
idna=2.7=py36_0
incremental=17.5.0=py36_0
intel-openmp=2018.0.3=0
ipykernel=4.8.2=py36_0
ipython=6.5.0=py36_0
ipython_genutils=0.2.0=py36_0
jedi=0.12.1=py36_0
jinja2=2.10=py36_0
jpeg=9b=h024ee3a_2
jsonschema=2.6.0=py36_0
jupyter_client=5.2.3=py36_0
jupyter_core=4.4.0=py36_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=7.2.0=hdf63c60_3
libgfortran-ng=7.2.0=hdf63c60_3
libpng=1.6.34=hb9fc6fc_0
libprotobuf=3.5.2=h6f1eeef_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=7.2.0=hdf63c60_3
libtiff=4.0.9=he85c1e1_1
markdown=2.6.11=py36_0
markupsafe=1.0=py36h14c3975_1
mistune=0.8.3=py36h14c3975_1
mkl=2018.0.3=1
mkl_fft=1.0.4=py36h4414c95_1
mkl_random=1.0.1=py36h4414c95_1
nb_conda=2.2.1=py36_0
nb_conda_kernels=2.1.0=py36_0
nbconvert=5.3.1=py36_0
nbformat=4.4.0=py36_0
ncurses=6.1=hf484d3e_0
ninja=1.8.2=py36h6bb024c_1
notebook=5.6.0=py36_0
numpy=1.15.0=py36h1b885b7_0
numpy-base=1.15.0=py36h3dfced4_0
olefile=0.45.1=py36_0
openssl=1.0.2o=h14c3975_1
pandoc=2.2.1=h629c226_0
pandocfilters=1.4.2=py36_1
parso=0.3.1=py36_0
pexpect=4.6.0=py36_0
pickleshare=0.7.4=py36_0
pillow=5.2.0=py36heded4f4_0
pip=10.0.1=py36_0
prometheus_client=0.3.1=py36_0
prompt_toolkit=1.0.15=py36_0
protobuf=3.5.2=py36hf484d3e_1
ptyprocess=0.6.0=py36_0
pyasn1=0.4.4=py36_0
pyasn1-modules=0.2.2=py36_0
pycparser=2.18=py36_1
pygments=2.2.0=py36_0
pyopenssl=18.0.0=py36_0
python=3.6.6=hc3d631a_0
python-dateutil=2.7.3=py36_0
pytorch=0.4.1=py36_cuda9.0.176_cudnn7.1.2_1
pyzmq=17.0.0=py36h14c3975_3
readline=7.0=ha6073c6_4
scikit-learn=0.19.1=py36hedc7406_0
scipy=1.1.0=py36hc49cb51_0
send2trash=1.5.0=py36_0
service_identity=17.0.0=py36h28b3542_0
setuptools=39.2.0=py36_0
simplegeneric=0.8.1=py36_2
six=1.11.0=py36_1
sqlite=3.24.0=h84994c4_0
tensorboard=1.9.0=py36hf484d3e_0
tensorflow=1.9.0=gpu_py36h02c5d5e_1
tensorflow-base=1.9.0=gpu_py36h6ecc378_0
tensorflow-gpu=1.9.0=hf154084_0
termcolor=1.1.0=py36_1
terminado=0.8.1=py36_1
testpath=0.3.1=py36_0
tk=8.6.7=hc745277_3
torchvision=0.2.1=py36_1
tornado=5.0.2=py36h14c3975_0
traitlets=4.3.2=py36_0
twisted=17.5.0=py36_0
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_1
werkzeug=0.14.1=py36_0
wheel=0.31.1=py36_0
xz=5.2.4=h14c3975_4
zeromq=4.2.5=hf484d3e_0
zlib=1.2.11=ha838bed_2
zope=1.0=py36_0
zope.interface=4.5.0=py36h14c3975_0
